# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AddQualifierIn": {
        "properties": {
          "is_hard_constraint": {
            "default": false,
            "title": "Is Hard Constraint",
            "type": "boolean"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "qualifier_effect": {
            "title": "Qualifier Effect",
            "type": "string"
          },
          "qualifier_type": {
            "title": "Qualifier Type",
            "type": "string"
          },
          "qualifier_value": {
            "title": "Qualifier Value",
            "type": "string"
          },
          "relation_id": {
            "title": "Relation Id",
            "type": "string"
          }
        },
        "required": [
          "relation_id",
          "qualifier_type",
          "qualifier_value",
          "qualifier_effect"
        ],
        "title": "AddQualifierIn",
        "type": "object"
      },
...
      "ConceptOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "augura_domain": {
            "title": "Augura Domain",
            "type": "string"
          },
          "canonical_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Unit"
          },
          "concept_name": {
            "title": "Concept Name",
            "type": "string"
          },
          "dq_column_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dq Column Role"
          },
          "layer": {
            "title": "Layer",
            "type": "integer"
          },
          "local_concept_id": {
            "title": "Local Concept Id",
            "type": "string"
          },
          "range_support_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Range Support Status"
          },
          "value_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Max"
          },
          "value_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Min"
          },
          "value_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Type"
          }
        },
        "required": [
          "local_concept_id",
          "concept_name",
          "augura_domain",
          "layer",
          "active"
        ],
        "title": "ConceptOut",
        "type": "object"
      },
...
      "DeactivateRelationIn": {
        "properties": {
          "relation_id": {
            "title": "Relation Id",
            "type": "string"
          }
        },
        "required": [
          "relation_id"
        ],
        "title": "DeactivateRelationIn",
        "type": "object"
      },
      "DirectRelationIn": {
        "description": "Relation légère proposée par le DAG (sans id pré-assigné).",
        "properties": {
          "default_strength": {
            "default": "moderate",
            "title": "Default Strength",
            "type": "string"
          },
          "mechanism_summary": {
            "default": "",
            "title": "Mechanism Summary",
            "type": "string"
          },
          "object_concept_id": {
            "title": "Object Concept Id",
            "type": "string"
          },
          "polarity": {
            "default": "neutral",
            "title": "Polarity",
            "type": "string"
          },
          "predicate": {
            "title": "Predicate",
            "type": "string"
          },
          "relation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relation Id"
          },
          "subject_concept_id": {
            "title": "Subject Concept Id",
            "type": "string"
          }
        },
        "required": [
          "subject_concept_id",
          "object_concept_id",
          "predicate"
        ],
        "title": "DirectRelationIn",
        "type": "object"
      },
...
      "EnrichApplyRequest": {
        "description": "Corps de POST /semantic/enrich/apply — un seul chemin renseigné à la fois.",
        "properties": {
          "add_qualifier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AddQualifierIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "deactivate_relation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeactivateRelationIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "direct_relations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DirectRelationIn"
            },
            "title": "Direct Relations",
            "type": "array"
          },
          "proposals": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposals"
          },
          "selected_concept_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Selected Concept Ids",
            "type": "array"
          },
          "selected_relation_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Selected Relation Ids",
            "type": "array"
          }
        },
        "title": "EnrichApplyRequest",
        "type": "object"
      },
      "EnrichApplyResponse": {
        "properties": {
          "concepts_added": {
            "default": 0,
            "title": "Concepts Added",
            "type": "integer"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "new_version": {
            "title": "New Version",
            "type": "string"
          },
          "previous_version": {
            "title": "Previous Version",
            "type": "string"
          },
          "relation_id_map": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Relation Id Map",
            "type": "object"
          },
          "relations_added": {
            "default": 0,
            "title": "Relations Added",
            "type": "integer"
          }
        },
        "required": [
          "new_version",
          "previous_version"
        ],
        "title": "EnrichApplyResponse",
        "type": "object"
      },
      "EnrichProposeAccepted": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "EnrichProposeAccepted",
        "type": "object"
      },
      "EnrichProposeRequest": {
        "properties": {
          "questions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PicotQuestionIn"
            },
            "title": "Questions",
            "type": "array"
          },
          "selected_concepts": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Selected Concepts",
            "type": "array"
          }
        },
        "title": "EnrichProposeRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PicotQuestionIn": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "picot": {
            "additionalProperties": true,
            "default": {},
            "title": "Picot",
            "type": "object"
          },
          "therapeutic_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Therapeutic Area"
          }
        },
        "required": [
          "id"
        ],
        "title": "PicotQuestionIn",
        "type": "object"
      },
...
      "RelationOut": {
        "description": "Relation causale de l'ontologie (B1) — sortie de GET /semantic/relations.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "default_strength": {
            "title": "Default Strength",
            "type": "string"
          },
          "mechanism_summary": {
            "title": "Mechanism Summary",
            "type": "string"
          },
          "object_concept_id": {
            "title": "Object Concept Id",
            "type": "string"
          },
          "polarity": {
            "title": "Polarity",
            "type": "string"
          },
          "predicate": {
            "title": "Predicate",
            "type": "string"
          },
          "relation_id": {
            "title": "Relation Id",
            "type": "string"
          },
          "subject_concept_id": {
            "title": "Subject Concept Id",
            "type": "string"
          }
        },
        "required": [
          "relation_id",
          "subject_concept_id",
          "predicate",
          "object_concept_id",
          "polarity",
          "default_strength",
          "mechanism_summary",
          "active"
        ],
        "title": "RelationOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "augura-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/semantic/concepts": {
      "get": {
        "operationId": "concepts_semantic_concepts_get",
        "parameters": [
          {
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "default": true,
              "title": "Active",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConceptOut"
                  },
                  "title": "Response Concepts Semantic Concepts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Concepts",
        "tags": [
          "semantic"
        ]
      }
    },
    "/semantic/enrich/apply": {
      "post": {
        "description": "Persiste un enrichissement dans l'ontologie GLOBALE (gated owner). Bump de version.",
        "operationId": "enrich_apply_semantic_enrich_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrich Apply",
        "tags": [
          "semantic"
        ]
      }
    },
    "/semantic/enrich/proposals/{job_id}": {
      "get": {
        "description": "Sert les propositions d'un job réussi (scopé tenant). Lues EN BASE\n(jobs.result_json), pas sur disque — relisibles depuis n'importe quel conteneur Modal.",
        "operationId": "enrich_proposals_semantic_enrich_proposals__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrich Proposals",
        "tags": [
          "semantic"
        ]
      }
    },
    "/semantic/enrich/propose": {
      "post": {
        "description": "Lance l'analyse de couverture + proposition LLM en job background (suivi par polling).",
        "operationId": "enrich_propose_semantic_enrich_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichProposeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichProposeAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrich Propose",
        "tags": [
          "semantic"
        ]
      }
    },
    "/semantic/relations": {
      "get": {
        "description": "Ontologie causale (B1). `?concept_id=` → sous-graphe (sujet ou objet = id).",
        "operationId": "relations_semantic_relations_get",
        "parameters": [
          {
            "in": "query",
            "name": "concept_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Concept Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RelationOut"
                  },
                  "title": "Response Relations Semantic Relations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Relations",
        "tags": [
          "semantic"
        ]
      }
    },
...
  }
}